+++ front/components/posts/cardPost/RightIcons.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useAxios } from '../../../hooks/useAxios';
 import { AxiosGetPostById } from '../../../interfaces';
+import { handleModal } from '../../../redux/slice/modalSlice';
 import { getPost } from '../../../redux/slice/postSlice';
 import { addRemoveLike, UserDataState } from '../../../redux/slice/userDataSlice';
 import { RootState } from '../../../redux/store';
@@ -13,26 +14,30 @@ const RightIcons = ({ id }: { id: number }) => {
 
   const [share, setShare] = useState(false);
 
-  const { likes } = useSelector<RootState, UserDataState>(state => state.userDataReducer);
+  const { likes, data } = useSelector<RootState, UserDataState>(state => state.userDataReducer);
 
   const api = useAxios();
 
   const handleLike = () => {
-    api
-      .post(`/post/${id}/like`)
-      .then(res => {
-        if (res.status === 200) {
-          dispatch(addRemoveLike({ post: { id } }));
-          api
-            .get(`/post/${id}`)
-            .then(({ data }: AxiosGetPostById) => {
-              const post = data.data.post;
-              dispatch(getPost(post));
-            })
-            .catch(err => console.log(err));
-        }
-      })
-      .catch(err => console.log(err));
+    if (!data?.access_token) {
+      dispatch(handleModal(true));
+    } else {
+      api
+        .post(`/post/${id}/like`)
+        .then(res => {
+          if (res.status === 200) {
+            dispatch(addRemoveLike({ post: { id } }));
+            api
+              .get(`/post/${id}`)
+              .then(({ data }: AxiosGetPostById) => {
+                const post = data.data.post;
+                dispatch(getPost(post));
+              })
+              .catch(err => console.log(err));
+          }
+        })
+        .catch(err => console.log(err));
+    }
   };
 
   return (
